feat(#302): Create separate classes for to_string and to_integer terms
