+++ test/factbase/terms/test_casting.rb
@@ -11,16 +11,6 @@ require_relative '../../../lib/factbase/term'
 # Copyright:: Copyright (c) 2024-2025 Yegor Bugayenko
 # License:: MIT
 class TestCasting < Factbase::Test
-  def test_to_str
-    t = Factbase::Term.new(:to_string, [Time.now])
-    assert_equal('String', t.evaluate(fact, [], Factbase.new).class.to_s)
-  end
-
-  def test_to_integer
-    t = Factbase::Term.new(:to_integer, [[42, 'hello']])
-    assert_equal('Integer', t.evaluate(fact, [], Factbase.new).class.to_s)
-  end
-
   def test_to_float
     t = Factbase::Term.new(:to_float, [[3.14, 'hello']])
     assert_equal('Float', t.evaluate(fact, [], Factbase.new).class.to_s)
